// src/dorc-web/src/apis/dorc-api/models/UserSearchResult.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface UserSearchResult
 */
export interface UserSearchResult {
    /**
     * @type {string}
     * @memberof UserSearchResult
     */
    DisplayName?: string | null;
    /**
     * @type {string}
     * @memberof UserSearchResult
     */
    FullLogonName?: string | null;
}